requirements.txt
attrs==25.4.0
Automat==25.4.16
babel==2.17.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
courlan==1.3.2
cryptography==46.0.3
cssselect==1.3.0
dateparser==1.2.2
defusedxml==0.7.1
filelock==3.20.0
greenlet==3.3.0
htmldate==1.9.4
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.12.2
itemloaders==1.3.2
jmespath==1.0.1
joblib==1.5.2
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
mysql-connector-python==9.5.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
parsel==1.10.0
playwright==1.56.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
pyee==13.0.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
queuelib==1.8.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
scikit-learn==1.7.2
scipy==1.15.3
Scrapy==2.13.4
scrapy-playwright==0.0.44
service-identity==24.2.0
six==1.17.0
threadpoolctl==3.6.0
tld==0.13.1
tldextract==5.3.0
tomli==2.3.0
trafilatura==2.0.0
Twisted==25.5.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
w3lib==2.3.1
zope.interface==8.1.1
